cookbook/integrations/flytekit_plugins/modin_examples/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    /Library/Developer/CommandLineTools/usr/bin/make requirements.txt
#
arrow==1.2.2
    # via jinja2-time
async-timeout==4.0.2
    # via redis
attrs==21.4.0
    # via
    #   jsonschema
    #   ray
binaryornot==0.4.4
    # via cookiecutter
certifi==2021.10.8
    # via requests
chardet==4.0.0
    # via binaryornot
charset-normalizer==2.0.12
    # via requests
checksumdir==1.2.0
    # via flytekit
click==8.1.1
    # via
    #   cookiecutter
    #   flytekit
    #   ray
cloudpickle==2.0.0
    # via flytekit
cookiecutter==1.7.3
    # via flytekit
croniter==1.3.4
    # via flytekit
cycler==0.11.0
    # via matplotlib
dataclasses-json==0.5.7
    # via flytekit
decorator==5.1.1
    # via retry
deprecated==1.2.13
    # via
    #   flytekit
    #   redis
diskcache==5.4.0
    # via flytekit
docker-image-py==0.1.12
    # via flytekit
docstring-parser==0.13
    # via flytekit
filelock==3.6.0
    # via ray
flyteidl==0.24.8
    # via flytekit
flytekit==0.31.0
    # via
    #   -r ../../../common/requirements-common.in
    #   flytekitplugins-modin
flytekitplugins-modin==0.31.0
    # via -r requirements.in
fonttools==4.31.2
    # via matplotlib
fsspec==2022.2.0
    # via
    #   flytekitplugins-modin
    #   modin
googleapis-common-protos==1.56.0
    # via flyteidl
grpcio==1.43.0
    # via
    #   flytekit
    #   ray
idna==3.3
    # via requests
importlib-metadata==4.11.3
    # via keyring
importlib-resources==5.6.0
    # via jsonschema
jinja2==3.1.1
    # via
    #   cookiecutter
    #   jinja2-time
jinja2-time==0.2.0
    # via cookiecutter
joblib==1.1.0
    # via scikit-learn
jsonschema==4.4.0
    # via ray
keyring==23.5.0
    # via flytekit
kiwisolver==1.4.2
    # via matplotlib
markupsafe==2.1.1
    # via jinja2
marshmallow==3.15.0
    # via
    #   dataclasses-json
    #   marshmallow-enum
    #   marshmallow-jsonschema
marshmallow-enum==1.5.1
    # via dataclasses-json
marshmallow-jsonschema==0.13.0
    # via flytekit
matplotlib==3.5.1
    # via -r ../../../common/requirements-common.in
modin==0.14.0
    # via flytekitplugins-modin
msgpack==1.0.3
    # via ray
mypy-extensions==0.4.3
    # via typing-inspect
natsort==8.1.0
    # via flytekit
numpy==1.22.3
    # via
    #   matplotlib
    #   modin
    #   pandas
    #   pyarrow
    #   ray
    #   scikit-learn
    #   scipy
packaging==21.3
    # via
    #   marshmallow
    #   matplotlib
    #   modin
    #   redis
pandas==1.4.1
    # via
    #   flytekit
    #   modin
pillow==9.0.1
    # via matplotlib
poyo==0.5.0
    # via cookiecutter
protobuf==3.19.4
    # via
    #   flyteidl
    #   flytekit
    #   googleapis-common-protos
    #   protoc-gen-swagger
    #   ray
protoc-gen-swagger==0.1.0
    # via flyteidl
py==1.11.0
    # via retry
pyarrow==6.0.1
    # via flytekit
pyparsing==3.0.7
    # via
    #   matplotlib
    #   packaging
pyrsistent==0.18.1
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   arrow
    #   croniter
    #   flytekit
    #   matplotlib
    #   pandas
python-json-logger==2.0.2
    # via flytekit
python-slugify==6.1.1
    # via cookiecutter
pytimeparse==1.1.8
    # via flytekit
pytz==2022.1
    # via
    #   flytekit
    #   pandas
pyyaml==6.0
    # via ray
ray==1.11.0
    # via flytekitplugins-modin
redis==4.2.0
    # via ray
regex==2022.3.15
    # via docker-image-py
requests==2.27.1
    # via
    #   cookiecutter
    #   flytekit
    #   responses
responses==0.20.0
    # via flytekit
retry==0.9.2
    # via flytekit
scikit-learn==1.0.2
    # via -r requirements.in
scipy==1.8.0
    # via scikit-learn
six==1.16.0
    # via
    #   cookiecutter
    #   grpcio
    #   python-dateutil
sortedcontainers==2.4.0
    # via flytekit
statsd==3.3.0
    # via flytekit
text-unidecode==1.3
    # via python-slugify
threadpoolctl==3.1.0
    # via scikit-learn
typing-extensions==4.1.1
    # via
    #   flytekit
    #   redis
    #   typing-inspect
typing-inspect==0.7.1
    # via dataclasses-json
urllib3==1.26.9
    # via
    #   flytekit
    #   requests
    #   responses
wheel==0.37.1
    # via
    #   -r ../../../common/requirements-common.in
    #   flytekit
wrapt==1.14.0
    # via
    #   deprecated
    #   flytekit
zipp==3.7.0
    # via
    #   importlib-metadata
    #   importlib-resources
